Add TimeProgressBar support for web
